Challange completed.
Note. I had a question mark on the top of the image when i insterted. I dont know why and how to solve it but Android studio helped and fixed it. If possible i want to know the possible reasons for that. I believe the image should be insterted in the right place otherwise we get warnings.